Add unit tests for maze generator
In this commit, we have added new tests for the maze-generator module. These tests are aimed at validating the correctness of the generated mazes under different conditions. They involve a wide array of scenarios including different maze dimensions and shapes (square, rectangular), as well as desired levels of randomness. In addition, the tests also evaluate how the seed parameter affects the maze generation process in order to guarantee reproducibility. The addition of these tests will help in ensuring that the algorithm performs as expected across different use cases.