Replace maze-generator.awk with mazy-mice.awk in tests
Changed calling of maze-generator.awk to mazy-mice.awk in mazy-mice tests. This is to reflect the renaming of maze-generator.awk to mazy-mice.awk to make names consistent. Also uncommented skip functionality in the tests to ensure they are run during testing process. This highlights that the script being tested matches the name of the test file, increasing code readability and maintainability.