Add new Book Store exercise
Introduced a new exercise "Book Store" which calculates the price of any conceivable shopping basket with books from a 5 book series, offering as big a discount as possible. The exercise includes an instruction guide, a set of test cases in test-book-store.bats, a solution template in book-store.awk, and corresponding meta-data. The goal is to encourage learners to improve their logical reasoning and problem-solving skills.