Add bowling exercise
This addition includes test definitions and judges for a new bowling exercise. The .awk file outputs a message prompting users to implement the solution. Also created a new config file, providing information about the exercise metadata. The test checks are conducted using the .bat file. Finally, instructions for the exercise are given in the instructions.md file.