Add new Book Store exercise (#254)
* Add new Book Store exercise

Introduced a new exercise "Book Store" which calculates the price of any conceivable shopping basket with books from a 5 book series, offering as big a discount as possible. The exercise includes an instruction guide, a set of test cases in test-book-store.bats, a solution template in book-store.awk, and corresponding meta-data. The goal is to encourage learners to improve their logical reasoning and problem-solving skills.

* Add source code for bats-support and bats-assert

The source code has been added as a new file for the bats-support and bats-assert libraries. The code has been concatenated and the comments have been removed to save space. All the necessary variables and functions required for the Bats test helpers are now available.