Add bowling exercise (#255)
* Add bowling exercise

This addition includes test definitions and judges for a new bowling exercise. The .awk file outputs a message prompting users to implement the solution. Also created a new config file, providing information about the exercise metadata. The test checks are conducted using the .bat file. Finally, instructions for the exercise are given in the instructions.md file.

* Add skip conditions to bowling game tests

Several tests in the bowling game have been updated to include a condition to skip the test if needed. This is important for test control and management, allowing us to prevent unnecessary or irrelevant tests from running and thus improve the efficiency of the overall testing process.