Add the "Proverb" exercise.
This exercise includes adding tests, example, instructions, and metadata for a new exercise called "Proverb". The main purpose of this exercise is to refine the learner's knowledge about manipulating and filtering strings in a list. The program takes an arbitrary list of inputs and generates a series of statements often found in a proverb. The final output will depend on these inputs and the order they've been placed in.