Add the "Proverb" exercise. (#256)
* Add the "Proverb" exercise.

This exercise includes adding tests, example, instructions, and metadata for a new exercise called "Proverb". The main purpose of this exercise is to refine the learner's knowledge about manipulating and filtering strings in a list. The program takes an arbitrary list of inputs and generates a series of statements often found in a proverb. The final output will depend on these inputs and the order they've been placed in.

* Reorder exercises in config.json

The 'Proverb' exercise has been reordered within the configuration file config.json. This change puts exercises in order according to their difficulty level. The difficulty level was also adjusted from 1 to 2 for 'Proverb'.

* Update author information in config.json files

The author data has been updated in the config.json files for both the 'automated-readability-index' and 'proverb' practice exercises. "Jegors Čemisovs" has been replaced with "rabestro" for the readability index, and "rabestro" has been added to the proverb exercise.

* Refactor the proverb exercise's example.awk file

The conditional statement has been simplified in the example.awk file for the 'proverb' exercise.