fix number of sentences in 3rd automated readability index test
It seems pretty clear that there are 14 sentences.  The example
implementation counted 13 because it requires whitespace after a period,
but one of the sentences has a closing quotation mark after the period.

The simplest solution seems to be to include the closing quotation mark
in the preceding sentence (and word).  Therefore we need to count its
length as part of RT and adjust how whitespace is handled a bit.

The resulting score necessarily goes down because the number of
sentences goes up.